Have CmManager keep a sorted copy of the Service list rather than re-creating
Having the manager object destroy and re-create the service list each time
can be problematic for application developers when the list and it's items
just dissapear from memory.
Add the required members and methods to CmService to store, set and compare
the order.
CmManager now sets order on CmService creation and updates it when the Service
list changes.