Keep a sorted list of services rather than recreating each time it changes
The service list was recreated each time the list exposed on the bus
changed. This was not only inconsistent with the behaviour of the other
lists contained in the manager but is also priving tricky for application
developers to maintain a consistent interface displaying the service list.

Switched to a model similar to connections and devices whereby we update
the existing list when the Services property changes.
Before emitting the services-changed signal, however, we sort the list.