CmManager: Add methods for finding GOobjects from their D-Bus object path
In order to be able to more intelligently keep track of all the connections,
devices and services the manager is aware of (such that there is only one
GObject representing each path on the D-Bus) we need to be able to search
for an object using the D-Bus path it represents.

Introduce 3 dumb iterating functions to do so.