CmService: Add order property and comparison function
Now that we intend to keep a managed list of services we need to be able
to sort it to reflect the order published on the bus.
Add an order property, and its getter, to CmService and a
cm_service_compare comparison function to compare the order of two
services.