CmService: Add method to make a service the default route
Now that we have a pointer to the creating manager we can get a pointer to the
first service in the list and move the current service before it to make said
service the default connection. In the situation where the service is not
connected we try to connect it first.
If either the connection or the move fails we return FALSE.